+++ src/error.rs
@@ -28,6 +28,12 @@ pub struct ErrDontCare;
 #[derive(Debug)]
 pub enum Error {
     IoError(io::Error),
+    /// Tried to create a texture with dimensions which are
+    /// greater than the maximum allowed texture size.
+    InvalidTextureSize {
+        width: u32,
+        height: u32,
+    },
     /// An error condition which is not further specified,
     /// will be slowly replaced by more useful error kinds.
     Unspecified,
@@ -43,6 +49,7 @@ impl From<ErrDontCare> for Error {
 #[derive(Debug)]
 pub enum LoadTextureError {
     IoError(io::Error),
+    InvalidTextureSize { width: u32, height: u32 },
     Unspecified,
 }
 
@@ -50,7 +57,36 @@ impl From<LoadTextureError> for Error {
     fn from(e: LoadTextureError) -> Self {
         match e {
             LoadTextureError::IoError(io) => Error::IoError(io),
+            LoadTextureError::InvalidTextureSize { width, height } => {
+                Error::InvalidTextureSize { width, height }
+            }
             LoadTextureError::Unspecified => Error::Unspecified,
+        }
+    }
+}
+
+/// An error returned by `Texture::new`.
+#[derive(Debug)]
+pub enum NewTextureError {
+    InvalidTextureSize { width: u32, height: u32 },
+}
+
+impl From<NewTextureError> for LoadTextureError {
+    fn from(e: NewTextureError) -> Self {
+        match e {
+            NewTextureError::InvalidTextureSize { width, height } => {
+                LoadTextureError::InvalidTextureSize { width, height }
+            }
+        }
+    }
+}
+
+impl From<NewTextureError> for Error {
+    fn from(e: NewTextureError) -> Self {
+        match e {
+            NewTextureError::InvalidTextureSize { width, height } => {
+                Error::InvalidTextureSize { width, height }
+            }
         }
     }
 }